requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.1
beautifultable==0.8.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2
Django==3.1.4
django-admin-honeypot==1.1.0
django-ajaximage==0.2.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-ip==1.0.2
django-jalali==4.0.0
django-js-asset==1.2.2
django-six==1.0.4
django-tinymce==3.2.0
django-widget-tweaks==1.4.8
fandogh-cli==1.24.1
gunicorn==20.0.4
idna==2.10
jdatetime==3.6.2
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
pyOpenSSL==19.1.0
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.25.11
websocket-client==0.57.0